+++ backend/src/main/java/com/forensic/repository/CaseAssignmentRepository.java
@@ -10,12 +10,12 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CaseAssignmentRepository extends JpaRepository<CaseAssignment, Long> {
-    List<CaseAssignment> findByCaseId(Integer caseId);
+    List<CaseAssignment> findByCaseId(Long caseId);
     
-    Optional<CaseAssignment> findByCaseIdAndUserId(Integer caseId, Long userId);
+    Optional<CaseAssignment> findByCaseIdAndUserId(Long caseId, Long userId);
     
     @Modifying
     @Transactional
     @Query("UPDATE CaseAssignment SET removedAt = CURRENT_TIMESTAMP WHERE caseId = :caseId AND userId = :userId AND removedAt IS NULL")
-    void revokeUserFromCaseTeam(Integer caseId, Long userId);
+    void revokeUserFromCaseTeam(Long caseId, Long userId);
 }
