+++ backend/src/main/java/com/forensic/repository/ClueRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-public interface ClueRepository extends JpaRepository<Clue, Integer> {
+public interface ClueRepository extends JpaRepository<Clue, Long> {
     Page<Clue> findAll(Pageable pageable);
 }
